refactor: remove React and rearrange files
