+++ entities/GameState.js
