--- src/App.css
@@ -1,7 +1,0 @@
-* {
-    width: 100%;
-}
-
-.app {
-    display: flex;
-}
