--- src/components/Player.css
@@ -1,4 +1,0 @@
-
-.player__avatar {
-    width: 50px;
-}
