--- src/components/PlayerList.css
@@ -1,4 +1,0 @@
-.player__list {
-    display: flex;
-    flex-direction: column;
-}
