--- src/components/board.css
@@ -1,133 +1,0 @@
-.board {
-  --width: 80px;
-  --height: 125px;
-  --length: calc(var(--width) * 9 + var(--height) * 2);
-  width: var(--length);
-  height: var(--length);
-  display: grid;
-  grid-template-rows: var(--height) repeat(9, var(--width)) var(--height);
-  grid-template-columns: var(--height) repeat(9, var(--width)) var(--height);
-  grid-template-areas: 't0 t1 t2 t3 t4 t5 t6 t7 t8 t9 t10' 't39 . . . . . . . . . t11' 't38 . . . . . . . . . t12' 't37 . . . . . . . . . t13' 't36 . . . . . . . . . t14' 't35 . . . . . . . . . t15' 't34 . . . . . . . . . t16' 't33 . . . . . . . . . t17' 't32 . . . . . . . . . t18' 't31 . . . . . . . . . t19' 't30 t29 t28 t27 t26 t25 t24 t23 t22 t21 t20'; }
-  .board__center {
-    grid-row: 2 / 11;
-    grid-column: 2 / 11; }
-
-.tile_0 {
-  grid-area: t0; }
-
-.tile_1 {
-  grid-area: t1; }
-
-.tile_2 {
-  grid-area: t2; }
-
-.tile_3 {
-  grid-area: t3; }
-
-.tile_4 {
-  grid-area: t4; }
-
-.tile_5 {
-  grid-area: t5; }
-
-.tile_6 {
-  grid-area: t6; }
-
-.tile_7 {
-  grid-area: t7; }
-
-.tile_8 {
-  grid-area: t8; }
-
-.tile_9 {
-  grid-area: t9; }
-
-.tile_10 {
-  grid-area: t10; }
-
-.tile_11 {
-  grid-area: t11; }
-
-.tile_12 {
-  grid-area: t12; }
-
-.tile_13 {
-  grid-area: t13; }
-
-.tile_14 {
-  grid-area: t14; }
-
-.tile_15 {
-  grid-area: t15; }
-
-.tile_16 {
-  grid-area: t16; }
-
-.tile_17 {
-  grid-area: t17; }
-
-.tile_18 {
-  grid-area: t18; }
-
-.tile_19 {
-  grid-area: t19; }
-
-.tile_20 {
-  grid-area: t20; }
-
-.tile_21 {
-  grid-area: t21; }
-
-.tile_22 {
-  grid-area: t22; }
-
-.tile_23 {
-  grid-area: t23; }
-
-.tile_24 {
-  grid-area: t24; }
-
-.tile_25 {
-  grid-area: t25; }
-
-.tile_26 {
-  grid-area: t26; }
-
-.tile_27 {
-  grid-area: t27; }
-
-.tile_28 {
-  grid-area: t28; }
-
-.tile_29 {
-  grid-area: t29; }
-
-.tile_30 {
-  grid-area: t30; }
-
-.tile_31 {
-  grid-area: t31; }
-
-.tile_32 {
-  grid-area: t32; }
-
-.tile_33 {
-  grid-area: t33; }
-
-.tile_34 {
-  grid-area: t34; }
-
-.tile_35 {
-  grid-area: t35; }
-
-.tile_36 {
-  grid-area: t36; }
-
-.tile_37 {
-  grid-area: t37; }
-
-.tile_38 {
-  grid-area: t38; }
-
-.tile_39 {
-  grid-area: t39; }
